Move reactor tests to a separate file
