+++ crates/tako/src/tests/mod.rs
@@ -1,3 +1,8 @@
+#[cfg(test)]
 mod integration;
+#[cfg(test)]
+mod test_reactor;
+#[cfg(test)]
 mod test_scheduler;
+
 pub mod utils;
